refactor: extract delivery review lifecycle
