+++ docs/02-delivery/engineering-epic-03/ticket-04-review-lifecycle-core-extraction.md
@@ -20,3 +20,10 @@ Extract one shared review core without reopening the Epic 02 boundary.
 - no PR-creation redesign
 - no storage-layout unification
 - no vendor-contract redesign
+
+## Rationale
+
+- Added `tools/delivery/review.ts` as the shared review lifecycle core for fetcher parsing, triager parsing, polling cadence, timeout handling, artifact persistence, thread-resolution persistence, and cumulative review outcome accumulation.
+- Rewired ticketed `poll-review`, ticketed `record-review`, and standalone `ai-review` to call that shared core through thin orchestrator adapters so the review lifecycle logic now lives in one place.
+- Kept PR-body refresh/rendering ownership in `tools/delivery/orchestrator.ts` so this ticket does not reopen the reviewer-facing metadata/rendering scope deferred to `E3.05`.
+- Preserved the existing artifact roots, timeout semantics, cumulative `patched` behavior, and standalone `needs_patch` to `operator_input_needed` mapping so Epic 02 review convergence behavior remains unchanged while the module boundary tightens.
